+++ modules/apps/rate-limiting/types/keys.go
@@ -2,6 +2,8 @@ package types
 
 import (
 	"encoding/binary"
+
+	errorsmod "cosmossdk.io/errors"
 )
 
 const (
@@ -30,7 +32,7 @@ var (
 	AddressWhitelistKeyPrefix = bytes("address-blacklist")
 	HourEpochKey              = bytes("hour-epoch")
 
-	PendingSendPacketChannelLength = 16
+	PendingSendPacketChannelLength = 64
 )
 
 // Get the rate limit byte key built from the denom and channelId
@@ -41,14 +43,18 @@ func RateLimitItemKey(denom string, channelID string) []byte {
 // Get the pending send packet key from the channel ID and sequence number
 // The channel ID must be fixed length to allow for extracting the underlying
 // values from a key
-func PendingSendPacketKey(channelID string, sequenceNumber uint64) []byte {
+func PendingSendPacketKey(channelID string, sequenceNumber uint64) ([]byte, error) {
+	if len(channelID) > PendingSendPacketChannelLength {
+		return nil, errorsmod.Wrapf(ErrInvalidChannelID, "channel %s with length %d is greater than the allowed length %d", channelID, len(channelID), PendingSendPacketChannelLength)
+	}
+
 	channelIDBz := make([]byte, PendingSendPacketChannelLength)
 	copy(channelIDBz, channelID)
 
 	sequenceNumberBz := make([]byte, 8)
 	binary.BigEndian.PutUint64(sequenceNumberBz, sequenceNumber)
 
-	return append(channelIDBz, sequenceNumberBz...)
+	return append(channelIDBz, sequenceNumberBz...), nil
 }
 
 // Get the whitelist path key from a sender and receiver address